move table view to workbookapp
